Wrap hardware and game settings up into a profile
Rebuilt the schema to have a concept of a profile.

Split all the tables up into separate schemas: profile, hardware and
game for better organization.

Rewrote dataaccess and service layer to be profile-centric. All settings
are per-user, so this makes sense, but it also will greatly simplify the
amnount of server coded needed for all the CRUD operations.